police-record-system (2)/police-record-system/requirements.txt
flask
flask-sqlalchemy
flask-login
flask-mail
flask-migrate
flask-wtf
python-dotenv
email-validator
rapidfuzz
indic-transliteration
pytest
